This is better. Make sure you use some infill, at least 25% otherwise the outer shell colapses partially and the outside surface is marred.
The thickness needs to be reduced again, probably by 0.1mm.
The diameter could also be reduced slightly.
The ball bearing falls out now, so definitely reduce its diameter, probably by 0.2mm.

Part of me wonders if the problems I was having before were still due to water
in the filament and now that so much more of it is gone, the prints are coming out
tighter and tighter, so all those adjustments need to be taken back...